refactor(auth): add authenticator
