+++ app/routes/api/auth/google.ts
@@ -1,6 +1,7 @@
 import type { ActionFunction, LoaderFunction } from '@remix-run/node';
 import { redirect } from '@remix-run/node';
-import { authenticator, PROVIDER_NAME } from '~/services/auth/google.server';
+import { authenticator } from '~/services/auth/index.server';
+import { PROVIDER_NAME } from '~/services/auth/google.server';
 
 export const action: ActionFunction = ({ request }) => {
   return authenticator.authenticate(PROVIDER_NAME, request);
