+++ app/routes/api/auth/google/callback.ts
@@ -1,5 +1,6 @@
 import type { LoaderFunction } from '@remix-run/node';
-import { authenticator, PROVIDER_NAME } from '~/services/auth/google.server';
+import { authenticator } from '~/services/auth/index.server';
+import { PROVIDER_NAME } from '~/services/auth/google.server';
 
 export let loader: LoaderFunction = ({ request }) => {
   return authenticator.authenticate(PROVIDER_NAME, request, {
