+++ app/services/auth/index.server.ts
@@ -1,0 +1,7 @@
+import type { User } from '@prisma/client';
+import { Authenticator } from 'remix-auth';
+import { sessionStorage } from '~/services/session/index.server';
+import getGoogleStrategy from './google.server';
+
+export const authenticator = new Authenticator<User>(sessionStorage);
+authenticator.use(getGoogleStrategy());
